Split up the items in interfaceItems
